requirements.txt
beautifulsoup4==4.11.1
bidict==0.22.0
bs4==0.0.1
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
eventlet==0.33.1
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.0.1
MarkupSafe==2.1.1
msal==1.18.0
pycparser==2.21
pyjade==4.0.0
PyJWT==2.4.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
ratelimit==2.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
uuid==1.30
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1